refactor(weread): isolate Web API backend
