+++ lib/WeReadWebApi/src/WeReadClient.h
