+++ lib/WeReadWebApi/src/WeReadStore.h
