+++ lib/WeReadWebApi/src/WeReadWebApi.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "WeReadBrowse.h"
+#include "WeReadClient.h"
+#include "WeReadHttpClient.h"
+#include "WeReadProtocol.h"
+#include "WeReadStore.h"
