+++ test/CMakeLists.txt
@@ -53,7 +53,7 @@ add_subdirectory(opds_filename)
 add_subdirectory(minibidi_arabic)
 add_subdirectory(combining_marks)
 add_subdirectory(time_utils)
-add_subdirectory(weread)
+add_subdirectory(weread_webapi)
 add_subdirectory(airpage)
 add_subdirectory(chinese_chess)
 add_subdirectory(sd_card_font_cache_format)
