+++ test/weread_webapi/CMakeLists.txt
@@ -1,13 +1,13 @@
 add_executable(WeReadProtocolTest
   WeReadProtocolTest.cpp
   ${REPO_ROOT}/lib/JsonParser/StreamingJsonParser.cpp
-  ${REPO_ROOT}/src/activities/apps/weread/WeReadProtocol.cpp
+  ${REPO_ROOT}/lib/WeReadWebApi/src/WeReadProtocol.cpp
 )
 
 target_include_directories(WeReadProtocolTest BEFORE PRIVATE
   ${crosspoint_simulator_SOURCE_DIR}/src
   ${REPO_ROOT}/src
-  ${REPO_ROOT}/src/activities/apps/weread
+  ${REPO_ROOT}/lib/WeReadWebApi/src
   ${REPO_ROOT}/lib/Epub
   ${REPO_ROOT}/lib/JsonParser
   ${REPO_ROOT}/lib/Logging
@@ -38,9 +38,9 @@ gtest_discover_tests(WeReadTouchGeometryTest)
 add_executable(WeReadStoreTest
   WeReadStoreTest.cpp
   ${REPO_ROOT}/lib/JsonParser/StreamingJsonParser.cpp
-  ${REPO_ROOT}/src/activities/apps/weread/WeReadBrowse.cpp
-  ${REPO_ROOT}/src/activities/apps/weread/WeReadProtocol.cpp
-  ${REPO_ROOT}/src/activities/apps/weread/WeReadStore.cpp
+  ${REPO_ROOT}/lib/WeReadWebApi/src/WeReadBrowse.cpp
+  ${REPO_ROOT}/lib/WeReadWebApi/src/WeReadProtocol.cpp
+  ${REPO_ROOT}/lib/WeReadWebApi/src/WeReadStore.cpp
   ${REPO_ROOT}/src/util/StringUtils.cpp
   ${REPO_ROOT}/lib/Utf8/Utf8.cpp
   ${crosspoint_simulator_SOURCE_DIR}/src/HalStorage.cpp
@@ -49,7 +49,7 @@ add_executable(WeReadStoreTest
 target_include_directories(WeReadStoreTest BEFORE PRIVATE
   ${crosspoint_simulator_SOURCE_DIR}/src
   ${REPO_ROOT}/src
-  ${REPO_ROOT}/src/activities/apps/weread
+  ${REPO_ROOT}/lib/WeReadWebApi/src
   ${REPO_ROOT}/lib/Epub
   ${REPO_ROOT}/lib/JsonParser
   ${REPO_ROOT}/lib/Logging
